requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv
aiohttp
numpy
pydantic>=2.0.0
python-multipart
requests
beautifulsoup4
markdownify
tqdm
markdown
setuptools